requirements.txt
python-telegram-bot
ethtoken
web3
